+++ src/f_cm_lightness.cpp
@@ -610,7 +610,7 @@ class TF_Spline : public TableFunction {
 public:
 	TF_Spline(const QVector<QPointF> *_points) {
 		std::vector<std::pair<float, float>> points(_points->size());
-		for(int i = 0; i < points.size(); ++i)
+		for(size_t i = 0; i < points.size(); ++i)
 			points[i] = std::pair<float, float>((*_points)[i].x(), (*_points)[i].y());
 		spline = new Spline_Calc(points);
 		_init(0.0, 1.0, 4096);
